LCS Overhaul, Floating Enemy and Header Bugfixing
-Allows for LCS to be triggered using "L" key.
-Floating enemy AI is dormant when in LCS.
-Floating enemy AI will become active if not in LCS.

+++ Thalassa/Thalassa/FloatingEnemy.cpp
@@ -1,8 +1,8 @@
 #include <crtdbg.h> // To check for memory leaks
 #include "AEEngine.h"
 #include "Structs.hpp"
-#include "utils.hpp"
 #include <iostream>
+#include <vector>
 #include "FloatingEnemy.hpp"
 
 //Constructor
@@ -15,7 +15,7 @@ Floatie::Floatie(float pos_X, float pos_Y, AEGfxVertexList* mesh) {
 
 void Floatie::CheckDistance(Player& player) {
 	//Get the distance between the player and the object
-	float distance = sqrt(pow(player.posX - gameobj.posX, 2) + pow(player.posY - gameobj.posY, 2));
+	float distance = (float)sqrt(pow(player.posX - gameobj.posX, 2) + pow(player.posY - gameobj.posY, 2));
 	if (distance <= 300) {
 		currentState = CHASE;
 	}
